refactor: move `errors` to `root` dir
