--- src/errors/index.ts
@@ -1,2 +1,0 @@
-export * from './FatcherError';
-export * from './isFatcherError';
