+++ src/index.ts
@@ -1,7 +1,6 @@
 export * from './helpers';
-export { FatcherError, isFatcherError } from './errors';
-import { combine } from './utils';
-export { combine };
+export { FatcherError } from './FatcherError';
+export { combine } from './utils';
 export * from './interfaces';
 export * from './fatcher';
 export * from './createScopedRequest';
